+++ integration_tests/__tests__/stack_trace-test.js
@@ -13,16 +13,16 @@ describe('Stack Trace', () => {
 
   it('prints a stack trace for runtime errors', () => {
     const result = runJest('stack_trace', ['runtime-error-test.js']);
-    const stdout = result.stdout.toString();
+    const stderr = result.stderr.toString();
 
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /1 test suite failed, 0 tests passed/
     );
     expect(result.status).toBe(1);
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /ReferenceError: thisIsARuntimeError is not defined/
     );
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /\s+at\s(?:.+?)\s\(__tests__\/runtime-error-test\.js/
     );
   });
@@ -32,29 +32,29 @@ describe('Stack Trace', () => {
       'runtime-error-test.js',
       '--noStackTrace',
     ]);
-    const stdout = result.stdout.toString();
+    const stderr = result.stderr.toString();
 
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /1 test suite failed, 0 tests passed/
     );
     expect(result.status).toBe(1);
 
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /ReferenceError: thisIsARuntimeError is not defined/
     );
-    expect(stdout).not.toMatch(
+    expect(stderr).not.toMatch(
       /\s+at\s(?:.+?)\s\(__tests__\/runtime-error-test\.js/
     );
   });
 
   it('prints a stack trace for matching errors', () => {
     const result = runJest('stack_trace', ['stack-trace-test.js']);
-    const stdout = result.stdout.toString();
+    const stderr = result.stderr.toString();
 
-    expect(stdout).toMatch(/1 test failed, 0 tests passed/);
+    expect(stderr).toMatch(/1 test failed, 0 tests passed/);
     expect(result.status).toBe(1);
 
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /\s+at\s(?:.+?)\s\(__tests__\/stack-trace-test\.js/
     );
   });
@@ -64,38 +64,38 @@ describe('Stack Trace', () => {
       'stack-trace-test.js',
       '--noStackTrace',
     ]);
-    const stdout = result.stdout.toString();
+    const stderr = result.stderr.toString();
 
-    expect(stdout).toMatch(/1 test failed, 0 tests passed/);
+    expect(stderr).toMatch(/1 test failed, 0 tests passed/);
     expect(result.status).toBe(1);
 
-    expect(stdout).not.toMatch(
+    expect(stderr).not.toMatch(
       /\s+at\s(?:.+?)\s\(__tests__\/stack-trace-test\.js/
     );
   });
 
   it('prints a stack trace for errors', () => {
     const result = runJest('stack_trace', ['test-error-test.js']);
-    const stdout = result.stdout.toString();
+    const stderr = result.stderr.toString();
 
-    expect(stdout).toMatch(/3 tests failed, 0 tests passed/);
+    expect(stderr).toMatch(/3 tests failed, 0 tests passed/);
     expect(result.status).toBe(1);
 
-    expect(stdout).toMatch(/Error: this is unexpected\./);
-    expect(stdout).toMatch(/this is a string\. thrown/);
+    expect(stderr).toMatch(/Error: this is unexpected\./);
+    expect(stderr).toMatch(/this is a string\. thrown/);
 
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /\s+at\s(?:.+?)\s\(__tests__\/test-error-test\.js/
     );
 
     // Make sure we show Jest's jest-resolve as part of the stack trace
     /* eslint-disable max-len */
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /Error: Cannot find module 'this-module-does-not-exist' from 'test-error-test\.js'/
     );
     /* eslint-enable max-len */
 
-    expect(stdout).toMatch(
+    expect(stderr).toMatch(
       /\s+at\s(?:.+?)\s\((?:.+?)jest-resolve\/build\/index\.js/
     );
   });
@@ -105,12 +105,12 @@ describe('Stack Trace', () => {
       'test-error-test.js',
       '--noStackTrace',
     ]);
-    const stdout = result.stdout.toString();
+    const stderr = result.stderr.toString();
 
-    expect(stdout).toMatch(/3 tests failed, 0 tests passed/);
+    expect(stderr).toMatch(/3 tests failed, 0 tests passed/);
     expect(result.status).toBe(1);
 
-    expect(stdout).not.toMatch(
+    expect(stderr).not.toMatch(
       /\s+at\s(?:.+?)\s\(__tests__\/test-error-test\.js/
     );
   });
