+++ packages/jest-cli/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "chalk": "^1.1.1",
     "graceful-fs": "^4.1.3",
+    "istanbul": "^0.4.4",
     "jest-changed-files": "^13.2.2",
     "jest-config": "^13.2.3",
     "jest-environment-jsdom": "^13.2.2",
@@ -14,9 +15,10 @@
     "jest-mock": "^13.2.2",
     "jest-resolve": "^13.2.2",
     "jest-runtime": "^13.2.3",
-    "jest-util": "^13.2.2",
     "jest-snapshot": "^13.2.3",
+    "jest-util": "^13.2.2",
     "json-stable-stringify": "^1.0.0",
+    "node-notifier": "^4.6.0",
     "sane": "^1.2.0",
     "which": "^1.1.1",
     "worker-farm": "^1.3.1",
