+++ packages/jest-cli/src/assets/jest_logo.png
