+++ packages/jest-config/package.json
@@ -9,15 +9,14 @@
   "main": "build/index.js",
   "dependencies": {
     "chalk": "^1.1.1",
-    "istanbul": "^0.4.2",
+    "istanbul": "^0.4.4",
     "jest-environment-jsdom": "^13.2.2",
     "jest-environment-node": "^13.2.2",
     "jest-jasmine2": "^13.2.3",
     "jest-mock": "^13.2.2",
     "jest-resolve": "^13.2.2",
     "jest-util": "^13.2.2",
-    "json-stable-stringify": "^1.0.0",
-    "node-notifier": "^4.6.0"
+    "json-stable-stringify": "^1.0.0"
   },
   "jest": {
     "rootDir": "./build",
